issue 13: file structure refactoring
