+++ .eslintignore
@@ -2,4 +2,4 @@
 node_modules/*
 dist/*
 webpackConfig/*
-src/stories/*
+src/ui-kit/storybook-introduction/*
