+++ src/ui-kit/image-card/image-card.tsx
